EE-4471 Respond with multiple category checks for multiple applicants
Note that this commit does enough to refactor the financial status request
and response
and get the existing tests working to make sure nothing is broken.  It does
not introduce any new functionality relating to the new request.

ALso note that this commit removes the income retrieval service which is
not being used anymore.  Better to remove it now than refactor the tests
just to remove it later.